update tests for configuration
